requirements.txt
fastapi
uvicorn
pyserial
requests
pydantic
piper-tts
